# docs/openapi.yaml
openapi: 3.1.0
info:
  title: SingAI API local
  version: 2.0.0
  description: |
    Contrato verificable de los tres servicios Express incluidos en el repositorio.
    La interfaz desplegada usa Firebase directamente; esas operaciones no son
    endpoints HTTP propios y por ello no forman parte de este contrato.
servers:
  - url: http://localhost:4100
    description: Servicio núcleo y API pública local
security: []
...
paths:
...
  /internal/email/magic-link:
    post:
      tags: [Interna - correo]
      summary: Entrega un enlace de acceso
      servers: [{ url: http://localhost:4102 }]
      security: [{ serviceKey: [] }]
      requestBody:
        $ref: '#/components/requestBodies/Mail'
      responses:
        '202': { description: Mensaje encolado o guardado localmente }
        '400': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '502': { $ref: '#/components/responses/Error' }
  /internal/email/verify-account:
    post:
      tags: [Interna - correo]
      summary: Entrega un enlace de verificación
      servers: [{ url: http://localhost:4102 }]
      security: [{ serviceKey: [] }]
      requestBody:
        $ref: '#/components/requestBodies/Mail'
      responses:
        '202': { description: Mensaje encolado o guardado localmente }
        '400': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '502': { $ref: '#/components/responses/Error' }
  /internal/email/password-reset:
    post:
      tags: [Interna - correo]
      summary: Entrega un enlace de restablecimiento
      servers: [{ url: http://localhost:4102 }]
      security: [{ serviceKey: [] }]
      requestBody:
        $ref: '#/components/requestBodies/Mail'
      responses:
        '202': { description: Mensaje encolado o guardado localmente }
        '400': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '502': { $ref: '#/components/responses/Error' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    serviceKey:
      type: apiKey
      in: header
      name: x-service-key
...
  requestBodies:
...
    Mail:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [to, link]
            properties:
              to: { type: string, format: email }
              name: { type: string }
              link: { type: string, format: uri }
  responses:
    Error:
      description: Error de validación, autorización o dependencia
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...